pyproject.toml
[tool.poetry]
name = "whatsappMessanger"
version = "0.1.0"
description = "Send automated messages via Selenium / WhatsApp Web to your friends"
authors = ["Florian Glaser <58706771+Floskinner@users.noreply.github.com>"]
license = "GNU General Public License v3.0"

[tool.poetry.dependencies]
python = "^3.9"
selenium = "^4.1.5"
pypng = "^0.0.21"
webdriver-manager = "^3.7.0"
PyQRCode = "^1.2.1"
pyzbar = "^0.1.9"
Pillow = "^9.1.1"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
pylint = "^2.13.9"
black = "^22.3.0"
pre-commit = "^2.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
